fix(impact): publish weights and ratios as JSON numbers, not strings
JacksonConfig quotes every BigDecimal, so holdingWeight, expectedImpact and
observedContribution were leaving as "0.303000" while contract 1.2 and the
contract's own example both show them unquoted. The frontend multiplies a
weight by 100 for display, which on a quoted string is concatenation or NaN -
and it fails silently, with no exception anywhere. Same defect in the linked
view's sentiment score and confidence, which contract 1.2 lists as numbers in
-1..1 and the frontend compares against thresholds.

Only valueImpact stays BigDecimal: money is a string on purpose so a figure a
lecturer verifies by hand cannot be rounded by a JavaScript double (SC-002).

The two endpoints each carried their own copy of the same entity-to-DTO
mapping, which is how one could have been fixed and the other left wrong.
Both now go through ImpactRowMapper.

Every existing test asserts on Java objects, so all 219 passed with the wrong
wire types. ImpactJsonShapeTest asserts on the serialised JSON instead, and
imports JacksonConfig explicitly - @WebMvcTest does not pick up a plain
@Configuration, so without it the slice would have used the stock ObjectMapper
and asserted the opposite of production behaviour while looking green.

+++ backend/src/main/java/com/fnpis/api/internal/dto/ImpactRow.java
@@ -16,19 +16,28 @@ import java.math.BigDecimal;
  * the price agreed. They are never blended into one score, because disagreement
  * is the informative case (CLAUDE.md domain notes).
  *
+ * <p><b>BigDecimal fields serialise as JSON strings, Double fields as numbers.</b>
+ * Same split as {@link HoldingRow}, enforced globally by {@code JacksonConfig}.
+ * Contract 1.2 puts money in strings and ratios, weights and percentages in
+ * numbers, so only {@code valueImpact} is BigDecimal here. Typing a weight as
+ * BigDecimal would emit {@code "0.303000"} with quotes, and the frontend's
+ * {@code weight * 100} would then be string concatenation or NaN.
+ *
  * @param holdingWeight the position's share of portfolio value at assessment
- *                      time, 0-1. Never null: a position that could not be
- *                      valued produces no row at all (EC-13) rather than a
- *                      zero-weight one that would read as "no impact"
+ *                      time, 0-1, as a display number. Never null: a position
+ *                      that could not be valued produces no row at all (EC-13)
+ *                      rather than a zero-weight one that would read as "no
+ *                      impact"
  * @param priceChangePct the session's move as a percentage, null when neither a
  *                      close pair nor a quote was available (EC-18). A percentage
  *                      rather than a ratio because it is display-facing
- * @param expectedImpact weight times sentiment score - what the news implied
+ * @param expectedImpact weight times sentiment score - what the news implied.
+ *                      A dimensionless ratio, so a number on the wire
  * @param observedContribution weight times the actual return - what happened.
  *                      Null whenever {@code priceChangePct} is
- * @param valueImpact   the money figure, serialized as a string so JavaScript's
- *                      double cannot round it (CLAUDE.md). Null when the return
- *                      is unknown
+ * @param valueImpact   the money figure, and the only string here, so
+ *                      JavaScript's double cannot round it (CLAUDE.md). Null
+ *                      when the return is unknown
  * @param alignment     INCONCLUSIVE covers both a sub-epsilon move and missing
  *                      price data - a real answer, not a failure to report
  */
@@ -36,10 +45,10 @@ import java.math.BigDecimal;
 public record ImpactRow(
         String symbol,
         String companyName,
-        BigDecimal holdingWeight,
+        Double holdingWeight,
         Double priceChangePct,
-        BigDecimal expectedImpact,
-        BigDecimal observedContribution,
+        Double expectedImpact,
+        Double observedContribution,
         BigDecimal valueImpact,
         Direction direction,
         Alignment alignment) {
